Add a read-receipt buffer for SMS relays in the event the read receipt is processed before the message
This buffer has a 15-entry capacity which should be more than enough to account for any out-of-sync messages. Oldest entries are removed first. This can be disabled by passing --disable-sms-read-buffer.
This introduces an additional property to Message as well, indicating whether the message has been read by the user. This is a more reliable source of truth as it takes into account the SMS read buffer.