Add improved logic for tracking the send progress of a message
When sending a message, iMessage has the following lifecycle:

finished: NO, sent: NO;  the message was literally just sent. nothing has happened to it yet.
finished: NO, sent: YES; the message was handed off to the daemon, still not really sent.
finished: YES, sent: NO; the message is finalized and preparing to send / being sent now. if errorCode != .noError, the message is failed and will not move on.
finished: YES, sent: YES; the message was sent successfully