Fix MMS failing to send on newer versions of Big Sur
CBInitializeFileTransfer will now automatically relocate attachments to the appropriate location within imagent, so any code that does this is now obsolete and should be removed. The underlying issue had to do with sandboxing and what paths imagent was allowed to access, as well as what conditions allowed imagent to access otherwise expected paths.